Modifications to the deployment scenarios
As the current scope of the document are the edge inside the
telecom operators network (before the radio) the definition of
Small edge is changed to represent the "more space available".

The following changes are added to Small edge
   - Changing physical sizes to metric system
   - Minimal depth is changed to 60 mm based on our feedback from customers

The  following changes added to all deployment scenarios:
   - Added witdh definition ot the physical size
   - Added front cooling, access and cabling
   - Added NEBS 3 compliancy
   - Smart NIC support added
   - Added a note about maintenance staff skillset

Note: In the future we should extend the scope of the document and
define an other deployment option for edge sites after the radio.

Change-Id: I342b8f028efcd1d865ba446b4258e9d84b58139b
Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>